requirement.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Authentication & OAuth
authlib==1.2.1
python-jose[cryptography]==3.3.0

# Database
pymongo==4.6.0

# Message Queue
pika==1.3.2

# HTTP Client & Session Management
httpx==0.25.2
starlette==0.27.0

# Data Validation & Models
pydantic==2.5.0
pydantic[email]==2.5.0

# Environment Variables
python-dotenv==1.0.0

# Date & Time handling (optional, but good to have)
python-multipart==0.0.6

# Additional utilities
typing-extensions==4.8.0